+++ src/racer/racer.zig
@@ -10,6 +10,7 @@ pub const Entity = @import("Entity/entity.zig");
 
 pub const Time = @import("Time.zig");
 pub const File = @import("File.zig");
+pub const Save = @import("Save.zig");
 
 pub const Random = @import("Random.zig");
 pub const Vector = @import("Vector.zig");